implemented the third endpoint
Implemented the third endpoint to add holidays of a certain country in a certain year to a users's calendar stored in the database. For it to work properly, user signup, login endpoints were implemented too. And to test it, there is also an endpoint that gets a user's calendar with all the events.